created spin_orbit files to specify functions related to the spin orbit problem. the dynamical_system lib is now more general
